refactor: move theme.js to assets-folder
