+++ src/components/containers/domains/social-media/SocialMediaButton.styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import theme from "../../../../theme";
+import theme from "../../../../assets/theme";
 
 const styles = StyleSheet.create({
     container: {
